--- db/connection.js
@@ -1,12 +1,0 @@
-const mysql = require("mysql2");
-require("dotenv").config();
-
-const connection = mysql.createConnection({
-  host: "localhost",
-  // MySQL username,
-  user: "root",
-  password: process.env.MYSQL_PASSWORD,
-  database: process.env.MYSQL_DB_NAME,
-});
-
-module.exports = connection;
