+++ db/schema.sql
@@ -1,14 +1,13 @@
-DROP DATABASE IF EXISTS emp_tracker_db;
-CREATE DATABASE emp_tracker_db;
+CREATE DATABASE employees_db;
 
-USE emp_tracker_db;
+USE employees_db;
 
-CREATE TABLE department(
+CREATE TABLE department (
     id INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
     dept_name VARCHAR(30) NOT NULL
 );
 
-CREATE TABLE emp_role(
+CREATE TABLE job_role (
     id INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
     title VARCHAR(30),
     salary DECIMAL(10,2),
@@ -16,13 +15,11 @@ CREATE TABLE emp_role(
     FOREIGN KEY (dept_id) REFERENCES department(id)
 );
 
-CREATE TABLE employee(
+CREATE TABLE employee (
     id INT NOT NULL AUTO_INCREMENT PRIMARY KEY,
     first_name VARCHAR(30),
     last_name VARCHAR(30),
     role_id INT NOT NULL,
-    manager_id INT,
-    FOREIGN KEY (manager_id) REFERENCES employee(id),
-    FOREIGN KEY (role_id) REFERENCES emp_role(id)
+    FOREIGN KEY (role_id) REFERENCES job_role(id),
+    manager_id INT REFERENCES employee(id)
 )
-
