+++ package.json
@@ -1,19 +1,92 @@
 {
-  "name": "hw",
+  "name": "company-db",
   "version": "1.0.0",
-  "description": "",
+  "description": "mock up application to track and update employees, departments, roles, and salaries. Made using node.js and mysql.",
   "main": "index.js",
+  "dependencies": {
+    "ansi-escapes": "^4.3.2",
+    "ansi-regex": "^5.0.1",
+    "ansi-styles": "^4.3.0",
+    "base64-js": "^1.5.1",
+    "bignumber.js": "^9.0.0",
+    "bl": "^4.1.0",
+    "buffer": "^5.7.1",
+    "chalk": "^4.1.2",
+    "chardet": "^0.7.0",
+    "cli-cursor": "^3.1.0",
+    "cli-spinners": "^2.6.1",
+    "cli-width": "^3.0.0",
+    "clone": "^1.0.4",
+    "color-convert": "^2.0.1",
+    "color-name": "^1.1.4",
+    "console.table": "^0.10.0",
+    "core-util-is": "^1.0.3",
+    "defaults": "^1.0.3",
+    "denque": "^2.0.1",
+    "dotenv": "^16.0.0",
+    "easy-table": "^1.1.0",
+    "emoji-regex": "^8.0.0",
+    "escape-string-regexp": "^1.0.5",
+    "external-editor": "^3.1.0",
+    "figures": "^3.2.0",
+    "generate-function": "^2.3.1",
+    "has-flag": "^4.0.0",
+    "iconv-lite": "^0.4.24",
+    "ieee754": "^1.2.1",
+    "inherits": "^2.0.4",
+    "inquirer": "^8.2.3",
+    "is-fullwidth-code-point": "^3.0.0",
+    "is-interactive": "^1.0.0",
+    "is-property": "^1.0.2",
+    "is-unicode-supported": "^0.1.0",
+    "isarray": "^1.0.0",
+    "lodash": "^4.17.21",
+    "log-symbols": "^4.1.0",
+    "long": "^4.0.0",
+    "lru-cache": "^6.0.0",
+    "mimic-fn": "^2.1.0",
+    "mute-stream": "^0.0.8",
+    "mysql": "^2.18.1",
+    "mysql2": "^2.3.3",
+    "named-placeholders": "^1.1.2",
+    "onetime": "^5.1.2",
+    "ora": "^5.4.1",
+    "os-tmpdir": "^1.0.2",
+    "process-nextick-args": "^2.0.1",
+    "pseudomap": "^1.0.2",
+    "readable-stream": "^3.6.0",
+    "restore-cursor": "^3.1.0",
+    "run-async": "^2.4.1",
+    "rxjs": "^7.5.5",
+    "safe-buffer": "^5.2.1",
+    "safer-buffer": "^2.1.2",
+    "seq-queue": "^0.0.5",
+    "signal-exit": "^3.0.7",
+    "sqlstring": "^2.3.3",
+    "string_decoder": "^1.3.0",
+    "string-width": "^4.2.3",
+    "strip-ansi": "^6.0.1",
+    "supports-color": "^7.2.0",
+    "through": "^2.3.8",
+    "tmp": "^0.0.33",
+    "tslib": "^2.4.0",
+    "type-fest": "^0.21.3",
+    "util-deprecate": "^1.0.2",
+    "wcwidth": "^1.0.1",
+    "yallist": "^4.0.0"
+  },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js"
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/brennan64/company-db.git"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "dependencies": {
-    "console.table": "^0.10.0",
-    "dotenv": "^16.0.0",
-    "inquirer": "^8.2.2",
-    "mysql2": "^2.3.3"
-  }
+  "bugs": {
+    "url": "https://github.com/brennan64/company-db/issues"
+  },
+  "homepage": "https://github.com/brennan64/company-db#readme"
 }
